requirements.txt
Flask>=2.3.0
mysql-connector-python>=8.0.0
reportlab>=4.0.0
gunicorn>=21.0.0
